+++ common/repositories/instance_list_repository.h
@@ -60,4 +60,25 @@ public:
 
 		return Strings::ToUnsignedInt(row[0]);
 	}
+
+	static int ShiftSuspendWindow(Database& db, uint16 instance_id, uint32_t suspended_at, uint32_t offline_seconds)
+	{
+		if (instance_id == 0 || suspended_at == 0 || offline_seconds == 0)
+		{
+			return 0;
+		}
+
+		auto results = db.QueryDatabase(fmt::format(
+			"UPDATE `{}` "
+			"SET `start_time` = `start_time` + {}, `expire_at` = `expire_at` + {} "
+			"WHERE `id` = {} AND `expire_at` > {}",
+			TableName(),
+			offline_seconds,
+			offline_seconds,
+			instance_id,
+			suspended_at
+		));
+
+		return results.Success() ? results.RowsAffected() : 0;
+	}
 };